feat: refactor adoptedTreeCard and adjust styling for skeleton
